move turbopack trace format into a trait
